requirements.txt
google-cloud-aiplatform>=1.25.0
google-cloud-speech>=2.20.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.7.0
PyYAML>=6.0
tenacity>=8.2.0
textstat>=0.7.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.20.0
rouge-score>=0.1.2
sentence-transformers>=2.2.0
nltk>=3.8
scikit-learn>=1.0.0